add mjs test
