+++ .github/workflows/test.yml
@@ -12,6 +12,7 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - run: npm install
+      - run: npm run build
       - run: npm test
 
   Lint:
