+++ README.md
@@ -4,6 +4,8 @@ Use eslint in the browser
 ```js
 import * as eslint from "eslint-linter-browserify";
 
+// or const eslint = require("eslint-linter-browserify");
+// or <script src="https://cdn.jsdelivr.net/npm/eslint-linter-browserify/linter.js"></script>
 // or <script src="https://cdn.jsdelivr.net/npm/eslint-linter-browserify/linter.min.js"></script>
 
 const linter = new eslint.Linter();
