+++ test.mjs
@@ -1,0 +1,4 @@
+import {Linter as mjsLinter} from "./linter.mjs";
+import {testLinter} from "./testLinter.js";
+
+testLinter("mjs", mjsLinter);
